New dynamic link for et3Response.notification.tribunal.template.id
