+++ src/main/resources/application.properties
@@ -59,7 +59,7 @@ referral.template.id=${REFERRAL_TEMPLATE_ID:f975c1cc-e102-472a-a6cb-84dffa0a6d42
 rejected.template.id=${REJECTED_TEMPLATE_ID:e480e011-ffeb-4adc-9682-8758cf206375}
 et1Serving.template.id=${ET1SERVING_TEMPLATE_ID:fa49291f-49cf-4330-8031-c05c101eb32a}
 et1Serving.respondent.template.id=${ET1SERVING_RESPONDENT_TEMPLATE_ID:05af8477-7313-4ef1-9a8e-6b2d9d128b50}
-et3Response.notification.tribunal.template.id=${ET3RESPONSE_NOTIFICATION_TRIBUNAL_TEMPLATE_ID:34e5ed0e-e39d-4ea7-acd7-751a1d7fb467}
+et3Response.notification.tribunal.template.id=${ET3RESPONSE_NOTIFICATION_TRIBUNAL_TEMPLATE_ID:d4ced02e-f381-491a-855d-950a283315c0}
 et3Notification.template.myhmcts.id=${ET3_NOTIFICATION_TEMPLATE_MYHMCTS_ID:6983b7cf-a492-46da-8eaf-9751191eb736}
 et3Notification.template.citizen.id=${ET3_NOTIFICATION_TEMPLATE_CITIZEN_ID:dd211d08-ba0a-4e9e-8502-31b4ed5515e9}
